+++ src/models/index.ts
@@ -4,4 +4,5 @@ export * from './base-mongo-object-infos.models';
 export * from './historic.models';
 export * from './lock-field.models';
 export * from './maps.models';
+export * from './update-many-at-once-options.models';
 export * from './update-many-query.models';