+++ src/mongo-read-stream.ts
@@ -5,6 +5,7 @@ import { MongoClient } from './client';
 import { ClassType } from './models/class-type.models';
 import { MongoUtils } from './mongo-utils';
 import { BaseMongoObject } from './models';
+import * as _ from 'lodash';
 
 export type PageConsumer<T> = (data: T[]) => Promise<void>;
 export type ItemConsumer<T> = (data: T) => Promise<void>;
@@ -59,10 +60,15 @@ export class MongoReadStream<U extends BaseMongoObject, L extends BaseMongoObjec
 
 	private lastId: string = null;
 	private cursor: Cursor<Partial<U | L>> = null;
+	private hasAlreadyAddedIdConditionOnce: boolean = false;
+
+	get query(): object {
+		return _.cloneDeep(this._query);
+	}
 
 	constructor(
 			private mongoClient: MongoClient<U, L>,
-			private query: object,
+			private _query: object,
 			private pageSize: number,
 			private projection: object = {},
 			private customType?: ClassType<Partial<U | L>>,
@@ -77,7 +83,9 @@ export class MongoReadStream<U extends BaseMongoObject, L extends BaseMongoObjec
 	 */
 	public async forEachPage(consumerFn: PageConsumer<Partial<U | L>>): Promise<void> {
 		return new Promise<void>((resolve, reject) => {
-			this.pipe(new PageConsumerWritable(this.pageSize, consumerFn))
+			this
+					.on('error', reject)
+					.pipe(new PageConsumerWritable(this.pageSize, consumerFn))
 					.on('error', reject)
 					.on('finish', resolve);
 		});
@@ -89,7 +97,9 @@ export class MongoReadStream<U extends BaseMongoObject, L extends BaseMongoObjec
 	 */
 	public async forEach(consumerFn: ItemConsumer<Partial<U | L>>): Promise<void> {
 		return new Promise<void>((resolve, reject) => {
-			this.pipe(new ItemConsumerWritable(consumerFn))
+			this
+				.on('error', reject)
+				.pipe(new ItemConsumerWritable(consumerFn))
 					.on('error', reject)
 					.on('finish', resolve);
 		});
@@ -103,24 +113,36 @@ export class MongoReadStream<U extends BaseMongoObject, L extends BaseMongoObjec
 	}
 
 	public async _read(size: number): Promise<void> {
-		if (!(this.cursor && await this.cursor.hasNext())) {
-			if (this.lastId) {
-				(this.query as any)['_id'] = { $gt: MongoUtils.oid(this.lastId) };
+		try {
+			if (!(this.cursor && await this.cursor.hasNext())) {
+				if (this.lastId) {
+					this._query['$and'] = this._query['$and'] || [];
+					const andConditions: object[] = (this._query as any)['$and'];
+					// avoid to add multiple time the _id condition
+					if (this.hasAlreadyAddedIdConditionOnce) {
+						andConditions[andConditions.length - 1]['_id']['$gt'] = MongoUtils.oid(this.lastId);
+					} else {
+						andConditions.push({ _id: { $gt: MongoUtils.oid(this.lastId) } });
+						this.hasAlreadyAddedIdConditionOnce = true;
+					}
+				}
+				if (this.customType) {
+					this.cursor = await this.mongoClient.findWithType(this._query, this.customType, 0, this.pageSize, { _id: 1 }, this.projection);
+				} else {
+					this.cursor = await this.mongoClient.find(this._query, 0, this.pageSize, { _id: 1 }, this.projection);
+				}
 			}
-			if (this.customType) {
-				this.cursor = await this.mongoClient.findWithType(this.query, this.customType, 0, this.pageSize, { _id: 1 }, this.projection);
-			} else {
-				this.cursor = await this.mongoClient.find(this.query, 0, this.pageSize, { _id: 1 }, this.projection);
+			let item = null;
+			if (await this.cursor.hasNext()) {
+				item = await this.cursor.next();
 			}
+			if (item) {
+				this.lastId = item._id;
+			}
+			this.push(item);
+		} catch (e) {
+			this.emit('error', e);
 		}
-		let item = null;
-		if (await this.cursor.hasNext()) {
-			item = await this.cursor.next();
-		}
-		if (item) {
-			this.lastId = item._id;
-		}
-		this.push(item);
 	}
 
 }