+++ test/tests/collection-exists.ts
@@ -1,10 +1,9 @@
 import { N9Log } from '@neo9/n9-node-log';
 import test, { Assertions } from 'ava';
-import * as mongodb from 'mongodb';
-import { AggregationCursor } from 'mongodb';
 
-import { MongoClient, MongoUtils } from '../../src';
+import { MongoClient } from '../../src';
 import { BaseMongoObject } from '../../src/models';
+import { init } from './fixtures/utils';
 
 class SampleType extends BaseMongoObject {
 	public test: string;
@@ -12,15 +11,7 @@ class SampleType extends BaseMongoObject {
 
 global.log = new N9Log('tests');
 
-test.before(async () => {
-	await MongoUtils.connect('mongodb://localhost:27017/test-n9-mongo-client');
-});
-
-test.after(async () => {
-	global.log.info(`DROP DB after tests OK`);
-	await (global.db as mongodb.Db).dropDatabase();
-	await MongoUtils.disconnect();
-});
+init(test);
 
 test('[EXISTS] Create collection and test existence', async (t: Assertions) => {
 	const mongoClient = new MongoClient('test-' + Date.now(), SampleType, null);