+++ test/tests/date-parser.ts
@@ -2,9 +2,9 @@ import { N9Log } from '@neo9/n9-node-log';
 import test, { Assertions } from 'ava';
 import { Transform } from 'class-transformer';
 import * as _ from 'lodash';
-import * as mongodb from 'mongodb';
-import { MongoClient, MongoUtils } from '../../src';
+import { MongoClient } from '../../src';
 import { BaseMongoObject } from '../../src/models';
+import { init } from './fixtures/utils';
 import * as DateParser from '../../src/transformers/date-parser.transformer';
 
 export class WithDateEntity extends BaseMongoObject {
@@ -18,15 +18,7 @@ export class WithDateAndNoTransformerEntity extends BaseMongoObject {
 
 global.log = new N9Log('tests').module('date-parser');
 
-test.before(async () => {
-	await MongoUtils.connect('mongodb://localhost:27017/test-n9-mongo-client');
-});
-
-test.after(async () => {
-	global.log.info(`DROP DB after tests OK`);
-	await (global.db as mongodb.Db).dropDatabase();
-	await MongoUtils.disconnect();
-});
+init(test);
 
 test('[DATE-PARSER] Insert&update entity with date', async (t: Assertions) => {
 	const entity: WithDateEntity = {