+++ test/tests/delete-one.ts
@@ -1,21 +1,13 @@
 import { N9Log } from '@neo9/n9-node-log';
 import test, { Assertions } from 'ava';
-import * as mongodb from 'mongodb';
 
-import { MongoClient, MongoUtils } from '../../src';
+import { MongoClient } from '../../src';
 import { BaseMongoObject } from '../../src/models';
+import { init } from './fixtures/utils';
 
 global.log = new N9Log('tests').module('issues');
 
-test.before(async () => {
-	await MongoUtils.connect('mongodb://localhost:27017/test-n9-mongo-client');
-});
-
-test.after(async () => {
-	global.log.info(`DROP DB after tests OK`);
-	await (global.db as mongodb.Db).dropDatabase();
-	await MongoUtils.disconnect();
-});
+init(test);
 
 test('[DELETE-ONE] Delete one by id', async (t: Assertions) => {
 	const mongoClient = new MongoClient('test-' + Date.now(), BaseMongoObject, BaseMongoObject);