+++ test/tests/lock-fields-from-none.ts
@@ -1,9 +1,9 @@
 import { N9Log } from '@neo9/n9-node-log';
 import test, { Assertions } from 'ava';
 import * as _ from 'lodash';
-import * as mongodb from 'mongodb';
-import { MongoClient, MongoUtils } from '../../src';
+import { MongoClient } from '../../src';
 import { BaseMongoObject } from '../../src/models';
+import { init } from './fixtures/utils';
 
 class SampleComplexType extends BaseMongoObject {
 	public property?: {
@@ -25,15 +25,7 @@ const getLockFieldsMongoClient = (keepHistoric: boolean = false) => {
 
 global.log = new N9Log('tests').module('lock-fields');
 
-test.before(async () => {
-	await MongoUtils.connect('mongodb://localhost:27017/test-n9-mongo-client');
-});
-
-test.after(async () => {
-	global.log.info(`DROP DB after tests OK`);
-	await (global.db as mongodb.Db).dropDatabase();
-	await MongoUtils.disconnect();
-});
+init(test);
 
 test('[LOCK-FIELDS] Update one from no value should be locked', async (t: Assertions) => {